Do not go back to L/RSYM and any L/RMOD mod has been pressed
Allows for better return to typing.  I could have elected to stay on the
L/RSYM layer, but I would have had to make this behaviour standard in combos.h
as well, and it would have taken a long time to edit all the discrete
positions.  I’ll have to see how this works out for me in the coming weeks.