Remove oneshot shift
I kept running into problems with it, and it was annoying to have to wait for
the cooldown when I actually wanted to type something.  It might take some
time for me to get used to the regular way to shift again, but I already do it
with key-bindings anyway.